Fix creating domains
Domains are created in the server state when identity endpoints
may not be present, as they added by keystone.clinet state.
As result using keystonerc file is not possible, sinse discovery
of identity endpoint will fail.
This patch adds endpoint override to domain create request so
it is succedded.

Change-Id: I2788b622f019ca03c55240a75e538f3d40a48df9
Related-Prod: PROD-19125